:art: Moved and refactored services for code cleanup
