+++ src/uActivityPub.Tests/ContentPublishPostHandlerTests.cs
@@ -8,7 +8,10 @@ using RichardSzalay.MockHttp;
 using uActivityPub.Data;
 using uActivityPub.Helpers;
 using uActivityPub.Models;
+using uActivityPub.Notifications.Handlers;
 using uActivityPub.Services;
+using uActivityPub.Services.ContentServices;
+using uActivityPub.Services.HelperServices;
 using uActivityPub.Tests.TestHelpers;
 using Umbraco.Cms.Core.Configuration.Models;
 using Umbraco.Cms.Core.Models;
