+++ src/uActivityPub/Composers/UActivityPubComposer.cs
@@ -6,7 +6,11 @@ using uActivityPub.Data;
 using uActivityPub.Data.Migrations;
 using uActivityPub.Helpers;
 using uActivityPub.Notifications;
+using uActivityPub.Notifications.Handlers;
 using uActivityPub.Services;
+using uActivityPub.Services.ActivityPubServices;
+using uActivityPub.Services.ContentServices;
+using uActivityPub.Services.HelperServices;
 using Umbraco.Cms.Core;
 using Umbraco.Cms.Core.Composing;
 using Umbraco.Cms.Core.DependencyInjection;
